+++ src/hooks/useChatHistoryHook.ts
@@ -1,7 +1,12 @@
-import { get, limitToLast, orderByKey, query, ref, remove } from 'firebase/database';
+import { get, limitToLast, onValue, orderByKey, query, ref, remove } from 'firebase/database';
 import { useEffect, useState } from 'react';
 import { database } from '../firebase/firebase';
+import useBrandStore from '../stores/useBrandStore';
+import useChatStore from '../stores/useChatStore';
+import useProductStore from '../stores/useProductsStore';
+import useUIStateStore from '../stores/useUIStateStore';
 import useUserStore from '../stores/useUserStore';
+import { ChatItem } from '../types/chatItem';
 
 type ChatHistory = {
   shareId: string;
@@ -15,7 +20,10 @@ const useChatHistoryHook = () => {
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState<null | string>(null);
   const { user } = useUserStore();
-
+  const {  setCurrentKeywords,currentChatId } = useChatStore();
+  const { setProducts } = useProductStore();
+  const { setShowChatBotAndRecommend} = useUIStateStore()
+  const { setBrands} = useBrandStore();
   useEffect(() => {
     const fetchChatHistory = async () => {
       if (!user?.uid) {
@@ -84,7 +92,63 @@ const useChatHistoryHook = () => {
     }
   };
 
-  return { chatHistory, loading, error, deleteChatHistory };
+  // Firebase에서 채팅 기록 불러오기
+ 
+    useEffect(() => {
+      if (user?.uid && currentChatId) {
+        const messagesRef = ref(database, `users/${user.uid}/chats/${currentChatId}/messages`);
+        const unsubscribe = onValue(messagesRef, (snapshot) => {
+          const chatItems: ChatItem[] = [];
+  
+          snapshot.forEach((messageSnapshot) => {
+            const messageId = messageSnapshot.key;
+            const message = messageSnapshot.val();
+            if (messageId && message) {
+              chatItems.push({
+                ...message,
+                id: messageId,
+              });
+            }
+          });
+  
+          chatItems.sort((a, b) => new Date(a.timestamp).getTime() - new Date(b.timestamp).getTime());
+  
+          if (chatItems.length > 0) {
+            setChatHistory(chatItems);
+            const lastChatItem = chatItems[chatItems.length - 1];
+            if (lastChatItem.products) {
+              setProducts(lastChatItem.products);
+              setShowChatBotAndRecommend(true);  // 제품이 있을 때만 추천 표시
+            } else {
+              setProducts([]);
+              setShowChatBotAndRecommend(false);  // 제품이 없으면 추천 숨김
+            }
+            if (lastChatItem.brands) {
+              setBrands(lastChatItem.brands);
+            }
+            setCurrentKeywords(lastChatItem.keywords);
+          } else {
+            setChatHistory([]);
+            setProducts([]);
+            setBrands([]);
+            setCurrentKeywords('');
+            setShowChatBotAndRecommend(false);// 채팅 기록이 없으면 추천 숨김
+          }
+        });
+  
+        return () => unsubscribe();
+      } else {
+        setChatHistory([]);
+        setProducts([]);
+        setBrands([]);
+        setCurrentKeywords('');
+        setShowChatBotAndRecommend(false);// 사용자나 채팅 ID가 없으면 추천 숨김
+      }
+    }, [user, currentChatId, setProducts, setBrands, setChatHistory, setCurrentKeywords]);
+
+
+
+  return { chatHistory, loading, error, deleteChatHistory};
 };
 
 export default useChatHistoryHook;
